app/requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
bcrypt==4.1.3
billiard==4.2.0
black==24.4.2
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
decorator==5.1.1
Django==4.2.8
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-phonenumber-field==7.2.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
environs==11.0.0
executing==2.0.1
gunicorn==22.0.0
ipython==8.25.0
jedi==0.19.1
kombu==5.3.7
marshmallow==3.21.2
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.12.36
platformdirs==4.2.1
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
